Refactor DB query result type parsing to output floats instead of ints

+++ server/db/index.js
@@ -4,9 +4,9 @@ const dbConfig = require("../../env/db.config");
 
 const pool = new Pool(dbConfig);
 
-types.setTypeParser(20, (val) => parseInt(val));
-types.setTypeParser(23, (val) => parseInt(val));
-types.setTypeParser(1700, (val) => parseInt(val));
+types.setTypeParser(20, (val) => parseFloat(val));
+types.setTypeParser(23, (val) => parseFloat(val));
+types.setTypeParser(1700, (val) => parseFloat(val));
 
 module.exports = {
   query: (qStr, qParams) => pool.query(qStr, qParams)
